refactor: Rearrangement of files
Made changes in file arrangement.
