--- src/components/toolbox/Input.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import ImageUpload from './ImageUpload';
-
-const Inputs = ({ uploadedImage, setUploadedImage }) => (
-  <div>
-    <ImageUpload uploadedImage={uploadedImage} setUploadedImage={setUploadedImage} />
-  </div>
-);
-
-export default Inputs;
