--- src/utils/uploadFrame.js
@@ -1,1 +1,0 @@
-// export uploadImage
